Make CapybaraHelpers requireable
Useful for testing extensions
